fix(chat): count completed replies as new unread activity
