+++ apps/core/src/routes/v1/chats/rooms/[id]/mute/delete.ts
@@ -13,11 +13,13 @@ import { chatRoomSchema } from "@/schemas/chat-room.schema";
 import {
   getChatRoomPinnedMessageCounts,
   getChatRoomSidebarFlags,
-  getChatRoomUnreadCounts,
-  getChatRoomUnreadMentionCounts,
   mapChatRoom,
   requireChatRoomUserAccess,
 } from "../../helpers";
+import {
+  getChatRoomUnreadCounts,
+  getChatRoomUnreadMentionCounts,
+} from "../../room-unread";
 
 const paramsSchema = z.object({
   id: z
