+++ apps/core/src/routes/v1/chats/rooms/[id]/star/post.ts
@@ -14,11 +14,13 @@ import { chatRoomSchema } from "@/schemas/chat-room.schema";
 import {
   getChatRoomPinnedMessageCounts,
   getChatRoomSidebarFlags,
-  getChatRoomUnreadCounts,
-  getChatRoomUnreadMentionCounts,
   mapChatRoom,
   requireChatRoomUserAccess,
 } from "../../helpers";
+import {
+  getChatRoomUnreadCounts,
+  getChatRoomUnreadMentionCounts,
+} from "../../room-unread";
 
 const paramsSchema = z.object({
   id: z
